+++ CHANGELOG.md
@@ -50,6 +50,7 @@ This release introduces the following breaking changes:
 - feat(sumologicschemaprocessor): add allowlist and denylist to nesting processor [#880]
 - feat(sumologicschemaprocessor) allow aggregating attributes with given name patterns [#871]
 - feat(sumologicschemaprocessor): add squashing single values in nesting processor [#881]
+- feat(sumologicschemaprocessor): report attributes as fields [#874]
 - feat(extension/sumologic): mark install_token as opaque [#882]
 
 [v0.68.0-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.67.0-sumo-0...v0.68.0-sumo-0
@@ -57,6 +58,7 @@ This release introduces the following breaking changes:
 [#880]: https://github.com/SumoLogic/sumologic-otel-collector/pull/880
 [#871]: https://github.com/SumoLogic/sumologic-otel-collector/pull/871
 [#881]: https://github.com/SumoLogic/sumologic-otel-collector/pull/881
+[#874]: https://github.com/SumoLogic/sumologic-otel-collector/pull/874
 [#882]: https://github.com/SumoLogic/sumologic-otel-collector/pull/882
 
 ### Changed